requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wrapt==1.12.1
